Reorganize server files and database connections
Server index has been moved to root file, server functions have
been seperated into their appropriate files, and all database
connections refer to a single config file.
