+++ .gitignore
@@ -5,11 +5,10 @@
 npm
 npm-debug.log
 node_modules/
-pics/
-pids
 *.pid
 logs
 *.log
+**config.js
 
 # Directory for instrumented libs generated by jscoverage/JSCover
 lib-cov
