+++ app/database/database.js
@@ -1,0 +1,4 @@
+const mysql = require('mysql');
+const connection = require('../../config.js');
+
+// do things
