+++ app/routers/router.js
@@ -1,0 +1,12 @@
+const db = require('../database/database.js');
+
+module.exports = {
+  readComments: () => {
+    // query database
+    // return promise
+  },
+  readCount: () => {
+    // query database
+    // return promise
+  },
+};
